Move files out form Painter folder
